# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "description": "API endpoints for managing hotel bookings",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get all payments",
                "operationId": "65567ad73dd4b0cc01c4856a6bb7e600",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentWithId"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "operationId": "3cb7bc2c2ad5498b4ad9521811750cc1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentPOST"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentWithId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            }
        },
        "/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get a payment by ID",
                "operationId": "a4207c19fecbebc99932a8348b19adf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentWithId"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Update a payment by ID",
                "operationId": "7f3042f886092282c90cf678fa8b0d78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentPUT"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentWithId"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete a payment by ID",
                "operationId": "6ee6ad54e3a047f39a06ffcb4cfc6b7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Payment deleted successfully"
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaymentPOST": {
                "title": "Payment for POST",
                "description": "Payment object",
                "required": [
                    "booking_id",
                    "amount",
                    "payment_date",
                    "status"
                ],
                "properties": {
                    "booking_id": {
                        "description": "ID of the booking associated with the payment",
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "description": "Amount of the payment",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "payment_date": {
                        "description": "Date of the payment",
                        "type": "string",
                        "example": "2024-02-13"
                    },
                    "status": {
                        "description": "Status of the payment",
                        "type": "string",
                        "enum": [
                            "completed",
                            "pending",
                            "failed"
                        ]
                    }
                },
                "type": "object"
            },
            "PaymentPUT": {
                "title": "Payment for PUT",
                "description": "Payment object",
                "properties": {
                    "booking_id": {
                        "description": "ID of the booking associated with the payment",
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "description": "Amount of the payment",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "payment_date": {
                        "description": "Date of the payment",
                        "type": "string",
                        "example": "2024-02-13"
                    },
                    "status": {
                        "description": "Status of the payment",
                        "type": "string",
                        "enum": [
                            "completed",
                            "pending",
                            "failed"
                        ]
                    }
                },
                "type": "object"
            },
            "PaymentWithId": {
                "title": "Payment with ID",
                "description": "Payment object",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the payment",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "booking_id": {
                        "description": "ID of the booking associated with the payment",
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "description": "Amount of the payment",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "payment_date": {
                        "description": "Date of the payment",
                        "type": "string",
                        "example": "2024-02-13"
                    },
                    "status": {
                        "description": "Status of the payment",
                        "type": "string",
                        "enum": [
                            "completed",
                            "pending",
                            "failed"
                        ]
                    },
                    "created_at": {
                        "description": "Timestamp when the payment was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-13 12:00:00"
                    },
                    "updated_at": {
                        "description": "Timestamp when the payment was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-13 12:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    }
}